requirements.txt
fastmcp>=2.0.0
httpx>=0.25.0
pydantic>=2.0.0
sqlalchemy>=2.0.0
aiosqlite>=0.19.0
python-dotenv>=1.0.0
flask>=3.0.0
flask-cors>=4.0.0
pyjwt>=2.8.0
cryptography>=41.0.0
requests>=2.31.0
openai>=1.0.0
pyyaml>=6.0
pytest>=7.4.0
pytest-asyncio>=0.21.0